pyproject.toml
[tool.poetry]
name = "pracsignal"
version = "0.1.0"
description = ""
authors = ["sds7629 <sds7629@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
